tlshd: GNUTTLS_CIPHER_* are not C macros
"#if defined" does not work on an enum. The original commit did
not generate a compilation error, but does not work when tested
on a system with a non-kTLS-enabled GnuTLS.

To enable compilation to work on platforms with an older GnuTLS
library, remove support for SM4 ciphers. The explicit switch in
tlshd_initialize_ktls() is meant as a fallback for when GnuTLS
is not built with kTLS support, so perhaps SM4 cipher support
is not necessary in tlshd.

Fixes: a38e1f830ae1 ("Initial commit")
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>