tlshd: Set -EIO instead of -EACCES on local error
In a subsequent patch, tlshd will be able to send a handshake
completion notification to the kernel that contains information
about whether a handshake failed because it was rejected by the
remote or it failed because of a local problem (misconfiguration
or a software bug).

Record that status in tlshd_completion_status.

As a clean-up, the atexit() logger is replaced with an in-line
equivalent.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>