tlshd: Get the remote's x.509 peer ID on the client side
Now that we have a richer mechanism for reporting the session status
after a handshake, we can pass the remote peer's identity back to the
kernel.

We could do this in tlshd_start_tls_handshake(), but it's almost
surely going to be a little different for PSK. So add it into the
client-side's x.509 certificate verification function.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>