tlshd: add 'keyrings' configuration option
Looking up keys is only possible if the 'read' permission has been
set. As we do not want anyone to read the contents of a key we
need to ensure that the handshake daemon has the appropriate
permissions to do so.

This patch adds a 'keyrings' configuration option to link keyrings
into the 'session' keyring for the tlshd daemon. That way tlshd
always has the ability to read them.

Signed-off-by: Hannes Reinecke <hare@suse.de>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>