tlshd: Generate the priorities string ourselves
The kernel should not have to care which TLS implementation tlshd is
using. The GnuTLS priorities string is specific to GnuTLS. (OpenSSL
has a priorities string with a different format, for example).

We also want to be able to generate the cipher list automatically,
which user space is better suited to do by examining both GnuTLS
headers and the kernel's headers.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>