psk: set resulting peerid after TLS handshake succeeded
For PSK the client sends a list of possible identities
(where each identity refers to a distinct PSK),
the server picks one and returns that one to the client.
So client and server only have a single, identical
identity, and that one refers to the PSK being used for
the TLS handshake.

Signed-off-by: Hannes Reinecke <hare@suse.de>