tlshd: set resulting peerid after TLS handshake succeeded
For PSK, the client sends a list of possible identities (where each
identity refers to a distinct PSK). The server picks one and returns
that identity to the client.

So client and server only have a single, identical identity, and
that one refers to the PSK being used for the TLS handshake.

Signed-off-by: Hannes Reinecke <hare@suse.de>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>