tlshd: Pass tlshd_parameters to the verification functions
I misunderstood the gnutls_session_get_ptr(3) API, possibly because
many code examples I've audited (ab)use it without context or
explanation.

When used it properly, the full set of handshake parameters can be
passed into the verification functions so the global variables can
be replaced.

Note that, because tlshd never called gnutls_session_set_ptr(3),
@hostname was always set to NULL. That means tlshd in ktls-utils-0.8
always skipped the hostname part of certificate verification.

The peername returned by DNS is fairly likely not to match the
SubjectName in the certificate in many circumstances. To prevent
breakage, we'll pass NULL explicitly to
gnutls_certificate_verify_peers3() for the moment until we can get a
real non-DNS-based hostname for certificate verification.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>