tlshd: Replace use of strcat(3) in tlshd_make_priorities_string()
Use of function strcat is deprecated because string lengths cannot
be limited. Consider strlcat() as an alternative. Also see CERT
STR07-C

We recently removed the libbsd-devel dependency, so strlcat(3) is a
bit of a challenge. Thus the goal here is to ensure that strcat(3)
is used in a safe fashion.

Reported-by: Parfait 10.2 (defect group #2037)
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>