tlshd: Revert b586f7d97795
Ben Hutchings observed that when tlshd handles the client side of a
handshake, it may get either an IP address or domain name from the
kernel, depending on how the target server was specified in the
mount attempt. The result of a reverse DNS lookup is used for
verifying the server's certificate, but that result is not
necessarily a secure piece of information.

The reason this reverse query was added was to enable SNI when
mounting NFS servers by IP address. Mounting by IP address is
needed, in particular, for pNFS and fs_locations results that
contain only IP addresses.

Reverting this commit temporarily breaks mounting by IP address with
TLS. However, recreating the server's certificate with a
subjectAltName field that lists the NFS server's IP address is
enough to enable NFS clients that mount by IP to securely verify the
server without using SNI.

Addresses Issue #22.

Fixes: b586f7d97795 ("tlshd: Reverse DNS lookup of peername")
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>