tlshd: fix init of privkey from keyring
When loading the private key with material from a keyring, tlshd did not
initialize the static tlshd_privkey. Instead, it accidentally filled in
a local copy. This lead to a failure when performing the handshake, with
a general error message of:

  tlshd[7823]: gnutls: No supported cipher suites have been found. (-87)

To fix this issue, simply supply the pointer to tlshd_privkey, like already
done when loading the certificate from the config file.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>