tlshd: Add support for chained certs
This patch adds support for chained certs (e.g. client/server .pem file which
includes an intermediate CA cert as well as the actual client/server cert).

Tested with a client/server NFS setup (using xprtsec=mtls) where only the root
CA is in the system trust store and both the client and server use chained
certificates with an intermediate CA as part of the cert.

Fixes: #44

Signed-off-by: David Härdeman <david@hardeman.nu>