Rework priority string setting for PSK
The patch "ktls: restrict hash functions to supported sizes"
did not work as intended, as it would reset the entire
string, not append to the existing one.
So rework the priority modification to store the common
priority, and modify it as needed during client initialisation.
Which has the benefit of displaying the final string
in the logging output, too.

Signed-off-by: Hannes Reinecke <hare@suse.de>