add option to skip server certificate check
This is useful for debugging.  I have an NFS server in which non production
builds will generate an x509 on the fly and send that to the client.  The
client has no way to have proper truststore to verify that certificate, so the
client must ignore it.  To enable no certificate checking, use the following
config...

   [authenticate.client]
   no_certificate_check = 1

Testing done to verify...
 * Tested with x509.truststore=...
 * Tested with no_certificate_check=0
 * Tested with no_certificate_check=1