configure.ac: allow users to enable libquic and quic handshake
This patch allows users to enable libquic and quic handshake by:

  # ./configure --enable-quic

Meanwhile, if the quic module is found by checking header file
/usr/include/linux/quic.h is installed, it will add src/libquic
to subdir in src/Makefile.am to build it.

LIBQUIC_CFLAGS and LIBQUIC_LIBS are defined to pass the compile
flags and libs into tlshd Makefile.am, and HAVE_QUIC is defined
to enable the quic handshake code in tlshd source code.

LT_INIT is required to enable shared libraries for libquic, and
libtoolize is added in autogen.sh to tell where to find LT_INIT.

Signed-off-by: Xin Long <lucien.xin@gmail.com>