libquic: add libquic directory in src for quic handshake
libquic will provides users with handshake and other advanced APIs
for in-kernel QUIC, the files in it includes:

connection.c, crypto.c, libquic.h are the source code of libquic,
and in crypto.c the code for session setup is quite similar to
that for tls/tcp in tlshd directory.

quic.h will be installed to $(includedir)/netinet to include these
advanced APIs exported from libuic to users.

quic.man provides the details about in-kernel quic usage, including
how libquic can be used for quic handshake.

libquic.pc.in is used to generate libquic.pc for pkg-config, and it
will be installed to $(libdir)/pkgconfig.

Signed-off-by: Xin Long <lucien.xin@gmail.com>