tlshd: include socket ip_proto in tlshd_handshake_parms
This patch adds the socket's ip_proto to tlshd_handshake_parms and
fills it in tlshd_genl_valid_handler().

This enhancement will allow the identification of QUIC handshakes
with ip_proto == IPPROTO_QUIC for the next patch. Additionally, it
sets the groundwork for future use in identifying DTLS handshakes
with ip_proto == IPPROTO_UDP.

Signed-off-by: Xin Long <lucien.xin@gmail.com>