tlshd: always link .nvme default keyring into the session
A common use case for tlshd is to authenticate TLS sessions for the nvme
subsystem. Currently, the user has to explicitly list a keyring (even
the defautl one) in the configuration file so that tlshd running
as daemon (started via systemd) to find any key.

Thus always link the default .nvme keyring into the current session,
which makes the daemon work out of the box for default configurations.

Signed-off-by: Daniel Wagner <wagi@monom.org>