tlshd: Pass ETIMEDOUT from gnutls to kernel
We've had some QE work that's created a condition (some types of connection
instability) where the handshake attempt has timed out.  When this happens,
tlshd sends EACESS back to the kernel.  However, the kernel may not be
expecting this error in the context of some NFS operations, for example:
writeback.  It can handle ETIMEDOUT, and we would like the kernel to
perform its normal hard/soft retry routines for this case to re-connect to
the server.

Add an error switch that clearly denotes the error paths we'd like
to send back to the kernel.  For SUNRPC, there are other insteresting
errors that might be included (see call_conenct_status() in
net/sunrpc/clnt.c), but are ommitted here because we don't have evidence of
them in the wild

Signed-off-by: Benjamin Coddington <bcodding@redhat.com>