tlshd: Preserve pcache during tlshd_gnutls_priority_init()
Valgrind reports:

==110975== Invalid read of size 4
==110975==    at 0x405B01: tlshd_gnutls_priority_init (ktls.c:422)
==110975==    by 0x4039D1: main (main.c:109)
==110975==  Address 0x534b1ec is 1,548 bytes inside a block of size 8,304 free'd
==110975==    at 0x4847B4C: free (vg_replace_malloc.c:989)
==110975==    by 0x405A02: tlshd_gnutls_priority_init (ktls.c:373)
==110975==    by 0x4039D1: main (main.c:109)
==110975==  Block was alloc'd at
==110975==    at 0x484C214: calloc (vg_replace_malloc.c:1675)
==110975==    by 0x48CBD1F: gnutls_priority_init (in /usr/lib64/libgnutls.so.30.37.1)
==110975==    by 0x4059DE: tlshd_gnutls_priority_init (ktls.c:366)
==110975==    by 0x4039D1: main (main.c:109)

The cipher list is allocated as part of the pcache. But
tlshd_gnutls_priority_init() releases the pcache immediately after
gnutls_priority_cipher_list() returns, then continues to access
the returned cipher list, which is now free memory.

Refactor tlshd_gnutls_priority_init() so that the lifetime of the
pcache is extended until tlshd is done with the cipher list array.

Refresh of fix-certificate-verification-error