tlshd: Add tag filter types
To start, define a filter type for each x.509 certificate field that
someone might want to match. These are defined in a table for easy
extensibility, and read into a hash table at start-up so they can be
looked up quickly by filter type name. (The filter type names will
eventually be exposed to users).

The filter types are introduced separately to give a sense of what
kind of matching can be done.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>