tlshd: Parse filter definitions
Administrators can define filters which specify a filter type (e.g.,
which x.509 field is to be checked) and filter arguments (e.g., what
content is to be matched).

A subsequent patch will add the ability to specify the tags which
are added to a TLS session based on filter matches.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>